# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Swagger for backend API service\nGet the Bearer token on the Authentication Service\nJSON Link: \u003ca href=/swagger/doc.json\u003edocs.json\u003c/a\u003e",
        "title": "Swagger for Authentication API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/login": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "parameters": [
                    {
                        "description": "Login",
                        "name": "Login",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Users"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/lib.OutputFormat"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "Data": {
                                            "$ref": "#/definitions/auth.Token"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/merchant/{merchant_id}/omzet": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Merchant"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "{1}",
                        "name": "merchant_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "2022-03",
                        "name": "bulan",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "20",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "1",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/lib.OutputFormat"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "Data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/merchants.MerchantOmzet"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/merchant/{merchant_id}/outlet/{outlet_id}/omzet": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Outlet"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "{1}",
                        "name": "merchant_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "{1}",
                        "name": "outlet_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "2022-03",
                        "name": "bulan",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "20",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "1",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/lib.OutputFormat"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "Data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/outlet.OutletOmzet"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/register": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "parameters": [
                    {
                        "description": "Register",
                        "name": "Register",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Users"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/lib.OutputFormat"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "Data": {
                                            "$ref": "#/definitions/auth.Token"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "auth.Token": {
            "type": "object",
            "properties": {
                "MerchantID": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "RefreshTokenString": {
                    "type": "string"
                },
                "TokenString": {
                    "type": "string"
                },
                "UserID": {
                    "type": "integer"
                }
            }
        },
        "lib.OutputFormat": {
            "type": "object",
            "properties": {
                "Code": {
                    "type": "string"
                },
                "Data": {},
                "Errors": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "properties": {
                            "Error": {
                                "type": "string"
                            },
                            "Field": {
                                "type": "string"
                            }
                        }
                    }
                },
                "Message": {
                    "type": "string"
                },
                "Success": {
                    "type": "boolean"
                }
            }
        },
        "merchants.MerchantOmzet": {
            "type": "object",
            "properties": {
                "Date": {
                    "type": "string"
                },
                "MerchantName": {
                    "type": "string"
                },
                "Omzet": {
                    "type": "number"
                }
            }
        },
        "model.Users": {
            "type": "object",
            "properties": {
                "Password": {
                    "type": "string"
                },
                "UserName": {
                    "type": "string"
                }
            }
        },
        "outlet.OutletOmzet": {
            "type": "object",
            "properties": {
                "Date": {
                    "type": "string"
                },
                "MerchantName": {
                    "type": "string"
                },
                "Omzet": {
                    "type": "number"
                },
                "OutletName": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}